(hail test_x00_init.py)
