+++ test/tests_3_ccm/test_300_init.py
@@ -1,6 +1,6 @@
 import pytest
 
 
-@pytest.mark.nlb
+@pytest.mark.ccm
 def test_cni_started(test, cni_started, logger):
     assert test["state"]["cni"]["started"] is True
