+++ test/tests_5_nodes_pool_resize/test_502_ccm_nodes_deletion.py
