+++ test/tests_5_nodes_pool_resize/test_503_ccm_nodes_csr_validation.py
